pyproject.toml
[tool.poetry]
name = "mutual-followers-analyzer"
version = "0.1.0"
description = "Telegram bot for analyzing mutual Instagram followers"
authors = ["Developer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
aiogram = "^3.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
httpx = "^0.26.0"
aiohttp = "^3.9.1"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

